se actualizan los cambios solicitados en el antrior PR
